Add files via upload
Added putty output file as both a .log and .out as requested. 
For this test GPIO2 and GPIO0 were connected together, such that the output of GPIO2 is supposed to trigger the GPIO0 interrupt on every falling edge. 
The output is verified to be working, as the interrupt gets triggered on only negative edges, it outputs to the console the low state after interrupt handled.